fix: review findings — types, deps stripping, continue template, torch_dtype
Code review fixes for DAG blackboard implementation:

1. Add dependency_interfaces to code_continue trajectory render
   (was missing — truncated subtasks lost dependency context)

2. Fix type annotations: dict[str, str] → dict[str, Any] for subtask
   dicts containing depends_on lists (lines 693, 727, 853)

3. Strip [depends: ...] from description text in parser — prevents
   literal bracket text from appearing in templates

4. Add dependency_interfaces section to code_continue.j2 template
   (matching code.j2 and code_retry.j2)

5. Update decompose prompts (both full and concise) to request
   [depends: ...] format — model confirmed to follow this format

6. Replace deprecated torch_dtype with dtype in from_pretrained calls
   (TransformersProvider and sakana_d2l activation extraction)

All checks pass: ruff clean, mypy 76 files, 329 tests.

+++ libs/inference/src/inference/transformers_provider.py
@@ -94,7 +94,7 @@ class TransformersProvider(InferenceProvider):
 
         self._model = AutoModelForCausalLM.from_pretrained(
             self._model_name,
-            torch_dtype=resolved_dtype,
+            dtype=resolved_dtype,
         )
         self._model.to(self._device)
         self._model.eval()
